01-ai-agents/07-panaversity-course/01_model_configuration/pyproject.toml
[project]
name = "01-model-configuration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
